+++ src/blocks/store/file.clj
@@ -196,7 +196,7 @@
         false))
 
 
-  store/BlockEnumerator
+  store/EnumerableStore
 
   (-enumerate
     [this]